Use the typescript compiler to type-check our Javascript
Add a tsconfig.json with allowJs: true and checkJS: true to enable
checking types in our code in vscode, and also add a run of 'tsc'
to test.sh to do the same checks in CI.

Add a few type annotations to the code, and also stop using exando
properties on the row elements - instead store strings in row.dataset,
and add a tunesById map. The typescript compiler doesn't like the
expando (at least out of the box), and its better avoided anyways.