Switch to using Ecmascript modules
Structure our code as an Ecmascript module, and import jquery as an
ecmascript module. To resolve jquery via npm and convert it into
ES module format, use "rollup". We make a bundle of all dependencies
(currently just jquery) and import that as a module - this provides
a better live-editing and debugging experience than making a bundle
of everything.